added y value to ribbon nodes and refactored to seperate controller
